clinic-home-visit-system/services/auth-service/requirements.txt
# Auth Service Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0
redis>=5.0.1
structlog>=24.1.0
prometheus-fastapi-instrumentator>=6.1.0
python-multipart>=0.0.6
httpx>=0.26.0
